requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
future==0.18.2
gunicorn==20.0.4
idna==2.8
kiwisolver==1.1.0
line-bot-sdk==1.15.0
lxml==4.5.0
matplotlib==3.1.3
numpy==1.18.1
Pillow==7.0.0
progressbar2==3.47.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-docx==0.8.10
python-utils==2.3.0
requests==2.22.0
six==1.14.0
urllib3==1.25.8
